+++ src/features/writer/storage/save-load-menu.tsx
@@ -12,11 +12,9 @@ import { ScrollArea } from '@/components/ui/scroll-area';
 import type { StoredEntry } from '@/lib/storage';
 import { isStorageFull } from '@/lib/storage';
 import type { JSONContent } from '@tiptap/react';
-import { useMutation, useQuery } from 'convex/react';
 import { Trash2Icon } from 'lucide-react';
 import React from 'react';
-import { api } from '../../../../convex/_generated/api';
-import type { Id } from '../../../../convex/_generated/dataModel';
+import GeneratedPostsList from './generated-posts-list';
 import LoadConfirmationDialog from './load-confirmation-dialog';
 import ReplaceEntryDialog from './replace-entry-dialog';
 
@@ -37,9 +35,6 @@ export default function SaveLoadMenu({
     title: string;
     content: string;
   } | null>(null);
-
-  const generatedPosts = useQuery(api.posts.list);
-  const removePost = useMutation(api.posts.remove);
 
   const loadEntries = React.useCallback(() => {
     const stored = localStorage.getItem('one-blog-entries');
@@ -89,11 +84,6 @@ export default function SaveLoadMenu({
     setSelectedEntry(null);
     setSelectedGeneratedPost(post);
     setLoadConfirmOpen(true);
-  };
-
-  const handleDeleteGeneratedPost = (e: React.MouseEvent, postId: string) => {
-    e.stopPropagation();
-    removePost({ id: postId as Id<'posts'> });
   };
 
   const handleDeleteEntry = (e: React.MouseEvent, entryId: string) => {
@@ -170,49 +160,7 @@ export default function SaveLoadMenu({
               </div>
             )}
           </ScrollArea>
-          {generatedPosts && generatedPosts.length > 0 && (
-            <>
-              <DropdownMenuSeparator />
-              <DropdownMenuLabel>
-                Generated Posts ({generatedPosts.length})
-              </DropdownMenuLabel>
-              <ScrollArea className="max-h-64">
-                <div className="grid gap-1 p-2">
-                  {generatedPosts.map((post) => (
-                    <DropdownMenuItem
-                      key={post._id}
-                      className="flex-col items-start gap-1 p-3"
-                      onSelect={() =>
-                        handleGeneratedPostClick({
-                          title: post.title,
-                          content: post.content,
-                        })
-                      }
-                    >
-                      <div className="flex w-full items-start justify-between gap-2">
-                        <span className="text-sm font-medium">
-                          {post.title}
-                        </span>
-                        <Button
-                          variant="ghost"
-                          size="icon"
-                          className="hover:text-destructive size-6 shrink-0 opacity-50 hover:opacity-100"
-                          onClick={(e) =>
-                            handleDeleteGeneratedPost(e, post._id)
-                          }
-                        >
-                          <Trash2Icon className="size-3.5" />
-                        </Button>
-                      </div>
-                      <span className="text-muted-foreground text-xs">
-                        {post.wordCount} words • {post.domain}
-                      </span>
-                    </DropdownMenuItem>
-                  ))}
-                </div>
-              </ScrollArea>
-            </>
-          )}
+          <GeneratedPostsList onSelect={handleGeneratedPostClick} />
         </DropdownMenuContent>
       </DropdownMenu>
