+++ tsconfig.app.json
@@ -25,8 +25,9 @@
 
     "baseUrl": ".",
     "paths": {
-      "@/*": ["./src/*"]
+      "@/*": ["./src/*"],
+      "@convex/*": ["./convex/*"]
     }
   },
-  "include": ["src"]
+  "include": ["src", "convex"]
 }
